pyproject.toml
[project]
name = "temporal-langgraph"
version = "0.1.0"
description = "LangGraph workflow orchestration using Temporal"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "temporalio>=1.12.0",
    "langgraph>=0.4.7",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "tavily-python>=0.5.0",
    "duckduckgo-search>=6.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
] 